Factor classpath resolution step out of application
It was previously done during application runtime, but this prooved to
be unfeasible, as the amount of dynamic (reflection-based) invocations
required became many enough to hamper system performance. As a
solution, application startup is divided into two steps, with there
first being a bootstrap step. After this step, during which UPPAAL is
located and a proper classpath resolved, the main application is
started in a child process with an appropriate classpath.